requirements_backup.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
black==26.3.1
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
datasets==4.8.4
dill==0.4.1
et_xmlfile==2.0.0
fastapi==0.135.2
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pypdf==6.9.2
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
reportlab==4.4.10
requests==2.33.1
rich==14.3.3
ruff==0.15.8
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
tabulate==0.10.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
